Do not send email notifications for unapproved posts
Rework send_notifications so that it does not send email notifications for unapproved posts.
Store notifications_sent on post to determine which posts need to be sent. Default to False
but set to True for all existing posts to prevent mass spamming. There is still a potential
issue here if the script has been turned off for a while and is re-enabled.